Relocate test methods in various test classes
The test methods in EmailTest.php, EnvTest.php, and FilterTest.php have been repositioned for better organization. It is important to note that no significant changes were made to the functionalities of these methods; they were simply moved to a different part in their respective files for a more streamlined and logical order.